linux: Add cmpxchg "emulator".
cmpxchg instruction is not supported on i386 cpu.

This is very naive, dirty and ugly "implementation" (read: hack)
which just works for needs of busybox userspace environment.
I'm not able to persuate gcc11 to not emit cmpxchg with lock prefix.

For now, it's just "works". But it's really terrible.

Patch also makes endbr32 instruction as a NOP instruction.
Multibyte nops are not supported since pentium 2 and it
can bring some headcache when you need to run some nice
binary compiled with gcc with CET support enabled.